backend/requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
email-validator==2.1.0

# Database
motor==3.3.2
pymongo==4.6.1
beanie==1.24.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
bcrypt==4.1.2

# Data & Finance (Phase 1 - Basic requirements)
yfinance==0.2.36
pandas==2.2.0
numpy==1.26.3
# pandas-ta will be added in Phase 4 when needed for technical indicators

# File Processing (Phase 2+)
openpyxl==3.1.2
xlrd==2.0.1
PyPDF2==3.0.1
pdfplumber==0.11.0
tabula-py==2.9.0

# Caching
redis[hiredis]==5.0.1
# aioredis is deprecated, redis 5.x has async support built-in

# HTTP Client
httpx==0.26.0

# Scheduling & Background Tasks (Phase 4+)
apscheduler==3.10.4

# Utilities
python-json-logger==2.0.7

# Development
pytest==8.0.0
pytest-asyncio==0.23.5
black==24.1.1
flake8==7.0.0
